kam3/web/requirements.txt
Flask
Flask-Session
gunicorn
python-dotenv
redis
bcrypt
pyjwt
flask_hal
requests